backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20

# Base de datos
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Autenticación - VERSIONES CORREGIDAS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Supabase
supabase==2.9.0

# Validación
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilidades
python-dotenv==1.0.1
pillow==11.0.0

# Email
resend==2.4.0